chore: move 2fa related code into it's own OtpForm component
