+++ src/client/modules/Settings/components/OtpForm/index.ts
@@ -1,0 +1,1 @@
+export { OtpForm } from './OtpForm';
